Extract conversation.js and several functions to user and utils
