+++ src/utils/common.js
@@ -1,3 +1,5 @@
+const _ = require('underscore');
+
 function strToHex(str) {
   let hex, i;
   // let str = "\u6f22\u5b57"; // "\u6f22\u5b57" === "漢字"
@@ -19,7 +21,80 @@ function hexToStr(hexString) {
   return str;
 }
 
+function getInt(s) {
+  return new Promise(function(resolve, reject) {
+    if (_.isNumber(s) && s >> 0 === s) {
+      return resolve(s);
+    }
+    let x = parseInt(s);
+    if (isNaN(x)) {
+      return reject("polis_fail_parse_int " + s);
+    }
+    resolve(x);
+  });
+}
+
+function getBool(s) {
+  return new Promise(function(resolve, reject) {
+    let type = typeof s;
+    if ("boolean" === type) {
+      return resolve(s);
+    }
+    if ("number" === type) {
+      if (s === 0) {
+        return resolve(false);
+      }
+      return resolve(true);
+    }
+    s = s.toLowerCase();
+    if (s === 't' || s === 'true' || s === 'on' || s === '1') {
+      return resolve(true);
+    } else if (s === 'f' || s === 'false' || s === 'off' || s === '0') {
+      return resolve(false);
+    }
+    reject("polis_fail_parse_boolean");
+  });
+}
+
+function getIntInRange(min, max) {
+  return function(s) {
+    return getInt(s).then(function(x) {
+      if (x < min || max < x) {
+        throw "polis_fail_parse_int_out_of_range";
+      }
+      return x;
+    });
+  };
+}
+
+function extractFromBody(req, name) {
+  if (!req.body) {
+    return void 0;
+  }
+  return req.body[name];
+}
+
+function extractFromCookie(req, name) {
+  if (!req.cookies) {
+    return void 0;
+  }
+  return req.cookies[name];
+}
+
+function extractFromHeader(req, name) {
+  if (!req.headers) {
+    return void 0;
+  }
+  return req.headers[name.toLowerCase()];
+}
+
 module.exports = {
   strToHex,
-  hexToStr
+  hexToStr,
+  getInt,
+  getBool,
+  getIntInRange,
+  extractFromBody,
+  extractFromCookie,
+  extractFromHeader
 };
